Melta alpha
Adds a new weapon type, canister weapons. Currently basically a splice of projectile and energy, in the future it will hopefully be atmos based.
The melta has old melta itemsprite, with held sprites from lasgun. It is currently a range-4 30-damage rapidfire weapon.
This is basically pre-balance, no idea how this will work in action, but the max-shots*damage stat is the same as the lascarbine.
Currently the only way to get it is through tech-crafting from silver, or adminspawn.
Fuel cans also unrefillable without VV.