inquisitorial changes
Adds the Inquisitorial Outpost Aecor. It is located on an active moon orbiting the planet. (Currently inaccessible, waiting for Wel Ard to help me with coder brain stuff but it's here.)
Changes it so the Inquisitor's job title is just Inquisitor instead of Ordo Helican Inquisitor, thematically and design-wise it looks cleaner.