Updated Bolt Pistol Round & Bolter Round.
Reduced AP value on Bolter slightly and increased damage. Same for bolt pistol. This is so the Astartes is more effective at killing mobs with base gun damage since they ignored AP.

Bolt Pistol has a significant damage boost but their stats are quite bad. Accuracy and fire delay specifically with the small magazine size. It's still lower damage then a Bolter but at least it's not a hot pile of shit.