Landmarks Update
Loot spawners are not highly complex as of yet but will act as an effective tool for mapping in future. Should allow coders to add items to the map without having to modify the map itself, instead they can just modify the Spawner Code - it can also allow easy changes to economy by changing the spawn chances.

Added a lot of equipment, supplies, items and weapons that weren’t previously accessible to players or only existed as manually placed items on the map. There are spawners for different parts of the map and with different qualities, I have some plans to add in spawners specific only to Xenos Temples and Heretic Fortresses once we add those.

Spawners are quite detailed but you can always DM me about it if you are doing mapping and need some help. A lot of the spawners are used for specific locations on the map, the Guardsmen Loot Spawners exist purely for Imperium Armories as an example. Make sure you always pair Ammo spawners with a Weapon spawner from the same category, last thing you want to do is add guns to the armory without any ammo for them.