Extra Stuff
Discovered an airless plating while searching the map further.

Added an additional locked house and a key - further encouraging people to think of the town as their home.

Some fixes for Voxcasters and Voxes.

Extra clothes on the map for Pilgrim characters. They're already so poor the least they deserve is some nice shoes and outfits to put on.