Mapping Update
The color of the snow is darker in the underground areas[Map 2] and brighter up top[Map 4].

Changed the shape of the town so it's less blocky. Further loot edits.

Outpost has a Dormitory now which comes with lockable doors, Guardsmen or other outpost staff can essentially have a personal room they can chill out in that is secure and private. Was a bit weird that nobody had places to sleep in and a giant communal sleeping area is a bit too military for my liking.